+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Expose `Text.Ascii.Internal` and `Text.Ascii.QQ`.
+* Add `Ixed` instance (and supporting type instances) for `AsciiText`.
 
 ## 1.0.0 -- 2021-02-07